[pjrt] Rewrite the transpose kernels, removing our dependence on Eigen
For an 8x8 uint32_t transpose, we had:
- 4x `vinsertf128 ymm, ymm, xmm`
- 4x `vperm2f128`.

These are very expensive instructions because they cross the 128-bit lane boundary.

Now, we have `8x vinsertf128` but crucially, the inserted operand now comes from memory. This is important because modern X86 HW can easily broadcast on load which means that `vinsertf128` turns into a blend instead of a shuffle.

We use the same trick for handling matrices which are smaller than the vector width to accelerate the transpose. We still require a cross lane step but we cut down all the other shuffles in half compared to SSE2.

While we are here, don't claim to support kernels which don't exist. This makes the transpose system choose unoptimized implementations.

PiperOrigin-RevId: 563153287